requirements.txt
sanic
kafka-python
pymongo
manage.py
